Removed legacy resources. Removed unused code.

+++ app/platforms/zazo_android/src/main/java/com/zazoapp/client/asr/dsp/LinearInterpolation.java
@@ -43,7 +43,7 @@ public class LinearInterpolation {
             return samples;
         }
 
-        int newLength=(int)Math.round(((float)samples.length/oldSampleRate*newSampleRate));
+        int newLength= Math.round(((float)samples.length/oldSampleRate*newSampleRate));
         float lengthMultiplier=(float)newLength/samples.length;
         short[] interpolatedSamples = new short[newLength];
